cpufreq: s7d dvfs bringup [1/1]
PD#SWPL-163073

Problem:
S7D Silicon Bringup - Kernel - Power management - DVFS

Solution:
dvfs bringup

Verify:
S7D

+++ drivers/cpufreq/aml_cpufreq_v2.h
@@ -14,8 +14,8 @@
 #define MAX_VOLTAGE	                   0x7fffffff
 
 struct opp_node {
-	int rate; //unit khz
-	int volt; //unit uv
+	unsigned long rate; //unit khz
+	unsigned long volt; //unit uv
 };
 
 /*this is dsu vote related structure read from dts*/
